llvmpipe: Texture cache in 4 ubytes instead of 4 floats.
This is more a short term experiment than a long term commitment, as we'll
need to support higher precision textures too, as this will all be
be replaced by runtime generated code.

With this change most Mesa demos fps increased around 10%. Not a huge
improvement, but not a negligible one either.

+++ src/gallium/drivers/llvmpipe/lp_tex_cache.h
@@ -61,7 +61,7 @@ union tex_tile_address {
 struct llvmpipe_cached_tex_tile
 {
    union tex_tile_address addr;
-   float color[TEX_TILE_SIZE][TEX_TILE_SIZE][4];
+   uint8_t color[TEX_TILE_SIZE][TEX_TILE_SIZE][4];
 };
 
 #define NUM_ENTRIES 50
